Moved Stencils into their own file
